k8s: disable node pool auto upgrades
GKE automatic node pool upgrades have taken down two testnets now.
Ideally our deployments would be resilient to these events, but we'd
prefer to minimize disruption to public testnets in the near-term.
This changes means we'll need to manually upgrade the node pools on
occasion, but we can schedule that as part of a regular testnet
deployment so the downtime is part of the scheduled window.

This change has already been deployed via CLI.