transaction: introduce an AuthHash for use in binding signatures
This is immediately motivated by not wanting to have messy debug output for
binding signature verification, but also fits with the conceptual division
between "authorizing" and "effecting" data, which now both have their own hash
type.

One rough edge in nomenclature this reveals (which we should sand down somehow)
is the distinction between `AuthorizingData` (the trait) and
`AuthorizationData` (the response from the custody protocol).  Unsure of the
best thing to do there.