fix: filter for valid addrs in testnet script
We use `cut` which does not have awesome CSV parsing,
so malformed lines can make it into the allocations file.
Here we filter for strings that at least look like
a penumbra address.

Tested locally, and using `xsv` instead of `cut` resolves, but would
require everyone to install that locally. Punting for now.

Refs #1876.