transaction: add AddressViews to TransactionView
Unlike in the design suggested in #2037, which suggested replacing every
`Address` by an `AddressView`, this simply sticks the address views into the
`TransactionView` as sidecar data. The reason is that the design suggested in
`ActionView` types, even when an address is visible, don't have an `Address`
directly, but instead have a `Note` or some other structure with an embedded
address.  So propagating the `View`ness would require potentially replicating
more structures.  In the meantime, before we decide if that's a good idea, this
at least conveys the idea that the address view should be part of a transaction
view.