testnets: add an ERC20-like 10e18 asset to Pasiphae allocations.
We updated the `Amount` type to be 128-bits, but stalled out on eliminating all
our previous uses of `u64`s, and on ensuring that we correctly handled 128-bit
amount values in the client logic.

This commit adds a `pusd` denom to the testnets (a fake USDC) with a 10e18
display denomination to force the issue, adds it to the base allocations
(including the test account), and updates the network integration test to send
20pusd (an amount of the base denom which is greater than u64::MAX).

To merge this PR and ship the Pasiphae allocations, we'll need to ensure that
we can actually correctly handle 128-bit amounts.