ci: test penumbra-web-assembly crate
Updates the GH workflows to test compilation of the new
`penumbra-web-assembly` crate to target `wasm32-unknown-unknown`.
Some small changes were made to the crate config, such as isolating it
from the parent workspace, and bumping the `edition` to 2021, same as
our other crates.

We preserve the `wasm-compat` dummy crate for now, since that passes a
CI build, and will serve as a reference point until we have CI passing
on `penumbra-web-assembly`; thereafter `wasm-compat` can be removed.

Also snipped out some unrelated CI configs for appveyor & travis, which
came in via the Zpoken PR, but aren't relevant for our CI system.