build: straighten out feature flags
Several of our crates failed to build under certain combinations of
feature flags. Stepped through those crates and adjusted the feature
logic where required, based on the output from `cargo-all-features` [0].

Regarding runtime for this CI check, the wall time isn't bad, especially
if we parallelize it with the other rust checks:

  command: cargo check-all-features --release
  warm cache: 1m39s
  cold cache: 12m4s

However, the disk usage is an order of magnitude higher:

  dir size of `target/`
  cargo build --release: ~3.4GB
  cargo check-all-features --release 25G(!)

Given that GHA has a hard limit on cache size of 10GB [1], we cannot add a
CI workflow to enforce this check. That's fine, we'll just run the
command opportunistically when editing feature flags.

[0] https://crates.io/crates/cargo-all-features
[1] https://docs.github.com/en/actions/using-workflows/caching-dependencies-to-speed-up-workflows#usage-limits-and-eviction-policy

Closes #2210.