docs: include wasm in rustdoc
Pulls the wasm crate into the parent workspace, so `cargo doc` and
associated scripting can find it.

In the process, I promoted the release profile setting of `opt-level=s`,
which optimizes for minimal binary size [0], from the wasm crate Cargo.toml
to the workspace. I compared output sizes (but not runtime metrics)
between the two profiles, and saw that pcli is smaller, but pd grew slightly.
The `.wasm` artifacts were indeed slightly smaller with opt-level=s.
In the future, we can further optimize with e.g. lto=true,
but I'm not doing that here because the goal is publishing docs for
wasm, and that's accomplsihed.

Closes #2275.

[0] https://doc.rust-lang.org/cargo/reference/profiles.html#opt-level
[1] https://rustwasm.github.io/book/reference/code-size.html#optimizing-builds-for-code-size