ci: relayer container build
Towards #465. Here we define a container image that wraps the upstream
cosmos/relayer image with the custom scripting for bootstrapping clients
on the Penumbra testnet and preview networks. No other chain configs are
part of this container image as of yet. The container image will be
publicly available as `ghcr.io/penumbra-zone/relayer`. It isn't yet
deployed to the cluster: that'll come next.