proto: enable grpc reflection
To do this, we have to save the `proto_descriptor.bin` generated by
Prost/Tonic.  It's currently 300KB and changes with every proto edit, so I
attempted to configure it to use git lfs.

Then, in the `pclientd` and `pd` binaries, we add a grpc-web enabled version of
the reflection service.