ci: configure cargo-release for version bumps
We plan to use the tool 'cargo-release' to manage version bumps.
As specified in the cargo-dist docs, we must choose between separate
versions (and therefore releases) of each package, or version all
packages within the workspace in lockstep. We choose the latter, which
is consistent with how we've been versioning our "releases" of testnets.

We hope to finetune the CI logic so that minor and patch version bumps
will be treated differently.