Retroactively remove tombstoned validators & their delegators from tally
This also refactors the emergency proposal pass logic into a common function on the state which can
be called from the action handler and the governance component. We call it in both places, because
now a validator tombstoning can result in a change to the proposal passing state for an emergency
proposal, and we should check immediately when the epoch rolls over (which is now triggered by the
slashing event).