Resolves #2085, preventing unintentional chain restarts
This PR writes to a key `chain/halted/N` where `N` is the number of times the chain has previously
halted, every time the chain halts. Then, upon initialization of the `App`, the key
`chain/halted/TOTAL_HALT_COUNT` is checked to determine if the chain has halted before without the
software having been updated. The constant `TOTAL_HALT_COUNT` defined separately in `app` and
`narsil` (which both reuse the `chain` crate) must be manually incremented in the code every time
the chain halts, in order for the new version of the software to pass over the block height of the
chain halt (and all previous halts); otherwise, the chain will not start up again. If
`TOTAL_HALT_COUNT` is incremented, synchronizing from genesis will not cause the chain to halt on
the previous halt block height, because the check in the `App` for the halt key
`chain/halted/TOTAL_HALT_COUNT` will fail, and the chain will continue to sync.