helm: convert rc -> deployment
The first draft of the helm chart used ReplicationControllers to deploy
the tm/pd pods. That was good enough when we were destroying on every
CI run, but now that we want more fine-grained control of behavior, such
as "update pd image version, keep the rest the same", Deployment is the
correct resource type for us. Using it will allow us to use tooling like:

    kubectl set image deployment/<d>  pd=ghcr.io/penumbra-zone/penumbra:<tag>

Which will come in handy for the "patch" version handling.