ci: handle patch version deploys gently
Adjusts the CI logic so that releasing a "patch" version
will update the version of pd running on the cluster, but not create a
new genesis. This allows us to push bugfixes to an already running
testnet, without requiring all participants to reset and rejoin.
Releasing a "minor" version, i.e. a new testnet, will still generate a
new genesis, consistent with past behavior.

In the process, converts the helm chart resources rc -> deployment.
The initial version of the helm chart used ReplicationControllers to deploy
the tm/pd pods. That was good enough when we were destroying on every
CI run, but now that we want more fine-grained control of behavior, such
as "update pd image version, keep the rest the same", Deployment is the
correct resource type for us. Using it will allow us to use tooling like:

    kubectl set image deployment/<d> pd=ghcr.io/penumbra-zone/penumbra:<tag>

to handle deploying "patch" versions differently. Similarly, we add a
"component" label so that we can query for the new deployments based on
fullnode/validator status, effectively excluding "metrics" from the list
of deployments, so we can edit the image spec specifically for pd.

This commit has a large diff on the primary ci script, but it's mostly
a superficial change, refactoring it into functions. There's one logic
change snuck in here: the `helm_uninstall` function now actually uses
`helm uninstall`, as was noted in a TODO. The rest is new logic relating
to version checking and soft rollout behavior.