dex: roll back "bid price"/"ask price" terminology
The bid price is the price that someone is willing to pay to buy; the ask price
is the price that someone is willing to pay to sell.  In other words, they
differ by the _direction of trading_ and are different prices.

That's not what the effective price is, it's just the price, inclusive of fees,
when trading from asset 1 to asset 2: `Delta_1 = Lambda_2 * effective_price`.
We want to search for low prices of asset 2 in terms of asset 1.  Its inverse
is just its inverse, not the price of a trade in the opposite direction, which
will be different, because gamma prices in a spread in both directions.

The only reason we might want the inverse instead is that in some parts of the
current filling code -- which we won't be keeping once we implement fill_route
-- we round in different directions, in order to ensure that when we have a
partial fill of the _trade input_, we completely consume the _position
reserves_, which requires changing which side of the formula we treat as the
"given" quantity and where we place the rounding error.