Use GATs in TCT incremental serialization API
This allows the implementor of AsyncRead/AsyncWrite/Read/Write to choose what iterator or stream is
returned, which is useful when using rusqlite, because rusqlite's iterators are `!Send`, so we need to
customize that bound when returning a `dyn Iterator` (i.e. it can't be hardcoded to `Send` in the
trait definition).