view,custody: make account_group_id be actually optional
It's marked as an optional field in the protos, but it wasn't treated as
optional by the rust view server.  This is annoying, since it makes it very
hard to use the view RPCs without already having the account group ID in hand.
We should let it be optional, since in every current implementation of the view
RPC, the account group ID is already implied by the endpoint (all current
implementations only support one account group).