ci: fix patch release workflow
The CI changes in #2387 were incomplete: they should have updated the
PENUMBRA_VERSION var with the value of `github.ref_name`, so that the
tag name is used to identify the container image. The default value of
referencing an input will fail if the workflow is triggered from a tag,
causing "main" to be deployed instead.

Although not strictly necessary, we modify the vars for the preview
deploy, as well, to maintain parity.