ci: disable relayer again =(
Shuts off the relayer again, because it's stuck in a crash loop. See
details in #2330. We should consider replacing the full relayer
deployment with a simpler CI check that tries to build the configs,
then reports an error and stays stopped if it fails. The current setup
will continue to retry forever, which junks up the chains with "init"
connections.

This reverts commit fbc80b7f6ff3e8d24fa4a919ca69d556811799de.