pclientd: fix envfilter setting, unbreaking bin
The envfilter setting pulled in via #2327 inadvertently broke the `pclientd`
binary, by preventing it from running if the env var RUST_LOG was not set (the
default behavior). The other bins like pd and narsild handled this error
appropriately, via `or_else` and setting a default of "info". Porting that
logic to pclientd resolves.