shielded-pool: extract crate out of `app`, `transaction`
Part of #2288.  This is a proof-of-possibility for the architecture described
in that issue.  Now, the `penumbra-shielded-pool` crate is self-contained, with
both the on-chain structures required to track the shielded pool, and the
transaction actions needed to interact with it.

In the future, a fair amount of functionality currently in the
`penumbra-crypto` crate could now be lifted up into this crate, for instance
the definitions of the spend and output proof statements.