ibc: refactor IBC message handling around a `MsgHandler` trait
This replaces a temporary vendored copy of the `ActionHandler` trait that was
used to avoid the orphan rules (the IBC message domain types are not locally
defined, so the handler trait must be).  The `MsgHandler` trait reflects that
we don't have a separate check_stateful, and paves the way for pulling more of
the IBC code out into ibc-async.