dao: split out component into subcrate.
This unlocks splitting out the `stake` (which currently depends on the dao
component, to handle funding stream rewards to the dao) and then `governance`
(which depends on the `stake` component as well).

While the DAO isn't a "full component" at present, in the sense that we don't
have begin/end block hooks for it, it does have chain state and associated
transaction actions, and we might want it to have queries, events, RPCs, etc.