ci: update rust cache action v1 -> v2
We've gone over the 10GB per-repo limit for workflow caches on GH.
GH will automatically evict caches not used for a week or longer, but
we've had a bunch of active PRs with plenty of rebases, and went over,
~22GB/10GB. Manually cleaned up old caches in the GitHub UI [1] to
unbreak workflow runs. This change brings us up to speed with the latest
release for the associated GH action helper, which has some updated
caching settings in the changelog. May need to follow up with
refinements to ensure that we don't hit the cache limit again.

N.B. GH does not allow us to purchase a cache increase per repo, which
is unfortunate.

[0] https://docs.github.com/en/actions/using-workflows/caching-dependencies-to-speed-up-workflows#usage-limits-and-eviction-policy
[1] https://github.com/penumbra-zone/penumbra/actions/caches