ci: compress genesis file during deployment
During ramp-up to Testnet 52 (#2415), we opportunistically updated the
Discord allocations, and encountered #1783 again. While we need a
durable solution to that problem, we also want a quick fix, so we turn
to tried-and-true file compression: we gzip the generated json file,
base64 encode it to sideload it into the cluster, then decompress it as
part of the initContainer bootstrapping logic during node creation.
This is a "good enough" solution to buy time for the near future,
ensuring that we can continue to update Discord allocations when we
want. The compression ratio is approximately 3x, which gives us some
headroom.