ci: compress genesis file during deployment, again
Follow-up to #2512. Here we remove the ConfigMap generation from the
Helm logic, opting instead to create/delete the resource manually via direct
use of `kubectl`. This removes the large base64-encoded file from the
Helm manifest, sidestepping the 1MB limit. We're still subject to a 1MB
limit on the ConfigMap object itself, but the gzipped-then-base64d data
is ~590Kb right now, so we've got some more time.