ci: be patient during configmap creation
Sprinkling "sleep"s around is bad form, but we're chafing against some
tooling problems with Helm, mixing in some raw kubectl commands that
helm resources then depend on. We'll pause a few seconds to ensure that
the k8s api agrees the configmap exists, since it's referenced in the
persistentvolumes that mount that info in to the pods.