dex: temporarily disable spill test
There's a problem with how the way we currently do filling interacts with the
way we check spill prices: we only get the effective price _after_ we execute
along the frontier, at which point it's too late to bail out and not execute
the price-limit-exceeding trades.

We can fix this later, the only impact for now is that the price limit may be
exceeded, giving a suboptimal route.