fixpoint: correct silent overflow bug
The bug was in this code:
```
        x1y1.checked_shl(128)
            .and_then(|acc| acc.checked_add(x0y1))
            .and_then(|acc| acc.checked_add(x1y0))
            .and_then(|acc| acc.checked_add(x0y0 >> 128))
            .map(U128x128)
```
The first call was intended to return None when there are nonzero bits in the
high part of the product. But that's not what checked_shl does: the check is on
the shift amount, not the shifted bits. Oops.

Longer-term, we should expand the `Error` type to have numeric errors, and
change the arithmetic operations to return `Result`s. For now, we can correct
this bug and see what other errors in our stack it was hiding.