dex: make SwapExecutions directed
I'm not sure we should land this for testnet 52; it was more complicated to
make these changes and it feels like the code is doing the wrong thing.  It's
not really clear to me why the execution traces should be stored as ambient
global state in the object store, rather than just being returned by the
methods that do execution. But that refactoring isn't something we should stuff
in before the deadline.