ci: use xl runner for macos builds
The release workflow added for #1961 takes about 90m to complete, most
of which is building binaries for macOS. We were previously using the
`macos-11` runner, which has 3 cores. Here we use `macos-12-xl`, which
has 12 cores [0]. Based on testing in a private repo, we can expect
build times for the macOS bins to fall from about 80m -> 30m. Now the
Windows builds take the longest (~50m), but it doesn't appear we have xl runner
options for those [1]. We could opt to add custom runners to our org
account, see details in [2].

[0] https://docs.github.com/en/billing/managing-billing-for-github-actions/about-billing-for-github-actions#per-minute-rates
[1] https://docs.github.com/en/actions/using-github-hosted-runners/about-github-hosted-runners#supported-runners-and-hardware-resources
[2] https://docs.github.com/en/actions/using-github-hosted-runners/using-larger-runners