crates: enforce using `v0_34::abci`
Now that `tendermint-rs` supports Tendermint `0.34.x` alongside
`0.37.x`, we have to make sure that we refer to the ABCI messages
defined in the versioned crates as opposed to importing the re-exports
from the root crates which by convention, point to the domain types for
the latest Tendermint version. As of `0.37.x`, this is not a problem
but it could cause confusing breakage in the future.