deploy: raise max peers for testnet
Makes the value configurable as an input var, and raises the maximum for
inbound and outbound peers 50 -> 75. Haven't found a lot of info about
best practices on these values, but there's an implicit ceiling on
number of open file descriptors permitted in security limits. Our GKE
nodes set nofile limits to 1048576, so we have more than enough
headroom to make this change.

Closes #1922.