ci: reorder docs build
The nightly build on the rustdocs sometimes fails, and we don't want
that to break deploys of unrelated docs targets, like the guide.
Ideally we'd have separate jobs for each deploy run in parallel, but
that would cause our CI cache consumption to spike, and we're already
struggling to stay under the 10GB per-repo limit.