ci: fix rustdoc build
A recent version of cargo nightly broke the docs build. This happens
sometimes. For now, to unbreak docs builds, I'm pinning an old version
of nightly known to be working. There's nothing special about this
version, I just picked one "a few weeks ago" and it worked, so I'm
committing it.

Also tweaks the JMT docs version post merge of #2628.