build: vendor protos via buf
Updates the protobuf build process to use `buf` to pull in the relevant
upstream proto definitions from BSR. We intentionally pin older versions
of the cosmos deps, cosmos-sdk & ibc, for compatibility. The most recent
version of cosmos/ibc doesn't integrate with our existing tooling, due
to introduction of new "extension" and "extension_signature" attributes
on the tendermint::vote type, but we'll tackle updating that later.

Only a single change was made to our proto compiler tool: excising a
manual import of a tendermint "query.rs" codegen file, which is no
longer exported from the sdk protos. Removing it did not break the
build, so we weren't using the functionality in that file. The rest of
the diff to proto-compiler is updating relpaths.

The proto specs now live at a top-level `proto/` dir, and alongside
the vendored protos. Those directories are used as inputs to the build.
Relevant docs have been updated.

Towards #2184.