ci: bounce preview deployment of grpcui
The grpcui deployment stays running, even after the backing grpc service
gets recreated. That means that the UI only displays the old services
from the previous deployment, which isn't what we want. Bounce it so it
rescans the server reflection-reported services and displays those in
the UI.

The grpcui help usage recommends setting `-connect-fail-fast=false` for
sidecar use cases like ours; might do that on the backend, in the infra
repo.