wip: terminate based on the actual price
This aims to be a better solution than the temporary fix we did last week. Roughly, the idea is:

- `fill_route` always makes progress, so it will return an exact price;
- we can compare that exact price against the price limit.

This prevents looping infinitely based on no-profit arbs that are mis-estimated
as being epsilon less than the price limit -- as soon as one of them actually
executes, the exact price will be computed and checked against the price limit,
terminating the process.