ci: bounce osiris in the deployment workflows
Osiris crashes when a testnet endpoint goes down, and recovers quite
gracefully, so this change isn't strictly necessary, but it prevents
the deployment from crashlooping, and also means the "restart" count per
deployment is meaningful info. I've been bouncing the testnet deployment
as part of the release procedures, so automated this step makes it
harder to forget.