docs: update deployment of bespoke devnet
Explains how to run a custom devnet for testing purposes. This is
something I do fairly frequently, but recently others have inquired what
the steps look like.

Took a pass through the cluster docs and fixed a broken logs URL, and
updated the IP information to match the recent migration to Traefik.