fix: resolves #2692, put ephemeral address in memo field
The problem with bug #2692 was:
1. In `TxCmd::exec` we generate an ephemeral address, but in
manipulating it set the `memo` variable (`Option<MemoPlaintext>`)
to `None` when there is no other memo text. Thus, the
ephemeral address is not used.
2. Then, while planning the transaction, we end up populating the
previously empty memo field with a default memo, which fills the
sender field with a dummy address that is not legible to the sender.