build: require protoc version 23.x
Our dev docs were badly out of date in terms of recommended version for
`protoc` and related tooling. Installed the latest and greatest, updated
the CI workflows, and clarified the docs. We now mandate a specific
major version of the protoc tooling, so we can ensure that

  1) protos have been regenerated in a CI check
  2) binary artifacts like the descriptor file don't flap between
     dev workstations

There's a bit of friction introduced here, in that folks will need to
update the tooling locally, but we've been ignoring that need to date
and paying the cost in confusion tokens, so I view the new workful as
a reasonable tradeoff.

Also regenerates all code emitted by the proto build process so we're up
to spec with the latest.

Refs #2736.