build: don't require a specific protoc version
Let's dial down the strictness on the protobuf codegen story.
Here we retain the docs changes from #2739, but drop the requirement for
a specific major version of `protoc`. Similarly, we remove the CI check
that fails if diffs are detected in the generated `proto_descriptor.bin`
file, because often those diffs are cosmetic, rather than substantive.
We preserve the CI check for ensuring there are no diff elsewhere in the
codegen outputs.

Closes #2736.