test: add pcli view tx to smoke test
Adds exercising of `pcli view tx` logic to the smoke test suite.
Reuses the existing `pcli send tx` test, ensuring that the initial send
can be viewed as raw JSON, and that JSON can be used to construct a
TransactionView. Also adds a memo string so we can sanity-check that the
TV is in fact legible.

Closes #2759.