relayer: only configure penumbra chains
We encountered a long (~20s) timeout on all `rly` commands, and traced
it to a context timeout on a bad URL for the bablyon testchain. We're
still focusing on preview <-> devnet functionality, so let's exclude the
external testnet chains from relayer configuration until we're ready.

Also includes some documentation updates so it's a bit easier to
understand how test the relayer interactively. Eventually we'll want all
this logic bottled up in integration tests, e.g. `interchaintest`, but
still working toward that goal.

Refs #2612.