test: begin unittest for tx id
Let's call this a failed attempt to use TDD: we expected the new
unittest to fail, given that we observed str -> proto -> str failures in
the view server, but evidently this naive test doesn't catch it.
Something else in the view server logic is screwing things up;
we suspect calls to `hex::encode` in particular.