spec: fix tilde rendering
The original LaTeX works for me locally outside of mdbook, so
I thought there must be an issue perhaps with mdbook-katex.

I've replicated using the same version of mdbook-katex that
CI is using that the `\tilde`s don't render, but the
`\widetilde`s _do_ render, so I've switched to those across the
protocol documentation.