fix(pd): drop unused snapshots
Trying to resolve a memory leak in pd. These manual drops are a first
pass, and appear to reduce memory consumption, but there's still a leak,
according to bytehound. We'll continue to investigate.

Includes a missed `to_proto` that's a nice to have, but likely doesn't-
constitute a fix for our problem.

Refs #2867.

Co-Authored-By: Henry de Valence <hdevalence@penumbralabs.xyz>