test: ensure consistent behavior in release and debug mode
For tests that involve invalid proof inputs, there are two ways
they can fail based on whether we are in release and debug mode.

In debug mode, there is upstream in ark-groth16 an assertion
that the circuit is satisifed prior to returning the proof. For
invalid inputs, the tests will panic at the proving stage.

In release mode, the proof will be constructed, but since it was
generated with invalid inputs, it will not verify. In that case,
we must panic if the return value of the verification method
was `false`, which corresponds to the expected behavior for tests
that operate on invalid inputs.