feat!(proto) remove optional fields in protos
The cosmos fork of gogoproto, used to generate golang code from BSR,
doesn't support "optional" fields in proto3 syntax. Rather than sort
that out, we're removing use of optional to unblock integration work on
interchaintest.

This commit updates the protos; still to come is the required Rust
changes for compatibility.

Refs #2933.