feat(proto)!: remove optional fields in protos
The cosmos fork of gogoproto, used to generate golang code from BSR,
doesn't support "optional" fields in proto3 syntax. Rather than sort
that out, we're removing use of optional to unblock integration work on
interchaintest.

For the most part, we can still represent optionality by special-casing
zero (for blockheights) and empty (for encrypted memos). The exception
is governance protos, which require use of a submessage type to
represent the optional "withdrawn" state on proposal outcomes.
H/t to @plaidfinch for describing the submessage pattern to me,
and implementing it on the governance changes.

Closes #2933.