governance: move DAO related view functions to app crate
The reason for this is that previously the `put_dao_transaction`
and `pending_dao_transactions` methods were defined in the `StateReadExt`
and `StateWriteExt` traits in `penumbra-governance`. However, they
require `Transaction` from `penumbra-transaction` which would result
in a circular dependency.

Instead, we make a tiny extension trait in the app crate with these two
functions.