ci: grow cluster 4 -> 8 nodes
This change is not yet applied; will apply it as part of the 61 release
(#3046), during maintenance window to upgrade gke versions. The
substance of this change is to grow the number of GKE nodes in the
cluster, increasing headroom for adding more Penumbra nodes per network
environment. We're effectively limited by the number of GKE nodes as to
how many Penumbra nodes can exist on a given network, due to
`allow_duplication_ip` being set to false in our Tendermint/CometBFT
config.