proto: split RPCs into per-component QueryServices
Closes #3083.

Still outstanding work to do:

- [ ] Split up the server-side implementations currently in `pd`.
- [ ] Resolve the governance issues.

On the latter point, I discovered that large parts of the governance code don't
have any RPCs defined, and instead rely on having knowledge of the Rust code to
layer together raw state accesses.  This is practically impossible for any
third party client to do, so it's not an acceptable solution.  The shim code
that enabled this was removed as part of the refactoring, so we'll either need
to define the relevant RPCs, or leave that functionality disabled until we do.