ci: update testnet public ips
Removes reserved IPv4 addresses that are no longer used.
For HTTPS services, we now use a single entry IP to a Traefik daemonset
to handle traffic for all the various endpoints [0].

Regenerates the public IPs for the P2P services and commits them to
version control. We do this as part of release prep for Testnet 61 [1],
building on the deploy overhaul described in [2].

[0] https://github.com/penumbra-zone/penumbra/issues/2341
[1] https://github.com/penumbra-zone/penumbra/issues/3046
[2] https://github.com/penumbra-zone/penumbra/issues/2983