ci: smooth over deploy bumps
Makes changes encountered while deploying Testnet 61 on the new deploy
logic for the first time:

  * fixes a YAML whitespace error on the testnet external IPs
  * make sure that strategy=recreate for metrics, otherwise
    config changes may encountered a failed concurrent bind on the pvc
  * also clean up jobs, which can get stuck if there are pvc errors

Made these changes locally and ran the deploy logic from my workstation
to finalize the Testnet 61 setup. Couldn't use the GHA in this scenario,
because of a chicken-or-egg problem: we need the change in the .0 tag,
but that tag was already pushed; we can't use a .1 tag because that only
modifies an existing deployment.

Refs #2983, #3046.