feat(pcli): support new ibc query logic
Previously pcli was only capable of looking up info on IBC clients,
connections, and channels if the user provided a unique identifier for
the resource in question. To facilitate IBC in practice, we want to
enable querying for lists of these values, so the end user (or
developer) can inspect the current options in a legible format.

All but the `channel` query have been updated. Submitting as-is
to ensure it lands in next testnet.

Closes #2937.

Co-authored-by: Ava Howell <ava@penumbralabs.xyz>